fix: Fix application lifecycle
