+++ docs/Application-Life-Cycle.md
@@ -1,0 +1,14 @@
+
+
+1. Http Request 
+2. Router { ctx.route, ctx.request.query, ctx.config, ctx.routes }
+3. Middleware pipeline
+4. Parameter Binding { ctx.parameters (raw value) }
+5. Custom Validator Creation
+6. Converter & Validation  { ctx.parameters (cleansed value) }
+7. Custom Authorization Creation
+8. Authorization
+9. Controller Creation
+10. Method Invocation
+11. Result Execution
+12. Http Response
