+++ packages/core/src/index.ts
@@ -1,12 +1,12 @@
 import { val } from "typedconverter"
 export { val }
-export { AuthorizeCallback, AuthorizeMiddleware, RoleField, updateRouteAccess, Authorizer } from "./authorization";
+export { AuthorizeCallback, checkAuthorize, RoleField, Authorizer } from "./authorization";
 export { HeaderPart, RequestPart, BindingDecorator, binder } from "./binder";
-export { invoke } from "./middleware-pipeline";
+export { invoke } from "./application-pipeline";
 export { response } from "./response";
-export { analyzeRoutes, generateRoutes, printAnalysis } from "./route-generator";
+export { generateRoutes } from "./route-generator";
+export { analyzeRoutes, printAnalysis } from "./route-analyzer";
 export { router } from "./router";
-export { ValidationMiddleware } from "./validator";
 export { Class, consoleLog, findFilesRecursive, getChildValue, hasKeyOf, isCustomClass } from "./common";
 export { AuthDecoratorImpl, authorize } from "./decorator.authorize";
 export { bind } from "./decorator.bind";
