+++ packages/plumier/test/integration/middleware/middleware.spec.ts
@@ -200,21 +200,6 @@ describe("Middleware", () => {
             expect(spy.mock.calls).toEqual([[1], [3], [5], [6], [4], [2]])
         })
 
-        it("Should able to access action parameter from middleware", async () => {
-            class AnimalController {
-                @route.get()
-                get(a: number, b: string, c: boolean) {
-                    return "Body"
-                }
-            }
-            const app = await fixture(AnimalController)
-                .use(new AssertParameterMiddleware([1, "1", true]))
-                .initialize()
-            await Supertest(app.callback())
-                .get("/animal/get?a=1&b=1&c=1")
-                .expect(200)
-        })
-
     })
 
     describe("Controller Middleware", () => {
@@ -327,21 +312,6 @@ describe("Middleware", () => {
             expect(spy.mock.calls).toEqual([[1], [3], [5], [6], [4], [2]])
         })
 
-        it("Should able to access action parameter from middleware", async () => {
-
-            @middleware.use(new AssertParameterMiddleware([1, "1", true]))
-            class AnimalController {
-                @route.get()
-                get(a: number, b: string, c: boolean) {
-                    return "Body"
-                }
-            }
-            const app = await fixture(AnimalController)
-                .initialize()
-            await Supertest(app.callback())
-                .get("/animal/get?a=1&b=1&c=1")
-                .expect(200)
-        })
     })
 
     describe("Action Middleware", () => {
@@ -452,22 +422,6 @@ describe("Middleware", () => {
                 .get("/animal/get")
                 .expect(200)
             expect(spy.mock.calls).toEqual([[1], [3], [5], [6], [4], [2]])
-        })
-
-        it("Should able to access action parameter from middleware", async () => {
-
-            class AnimalController {
-                @middleware.use(new AssertParameterMiddleware([1, "1", true]))
-                @route.get()
-                get(a: number, b: string, c: boolean) {
-                    return "Body"
-                }
-            }
-            const app = await fixture(AnimalController)
-                .initialize()
-            await Supertest(app.callback())
-                .get("/animal/get?a=1&b=1&c=1")
-                .expect(200)
         })
     })
